requirements.txt
# Core dependencies with flexible versions for Python 3.13 compatibility
aiohttp>=3.9.0
aiosignal>=1.3.0
annotated-types>=0.6.0
anyio>=4.0.0
asgiref>=3.7.0
async-timeout>=4.0.0
attrs>=23.0.0
backoff>=2.2.0
bcrypt>=4.1.0
beautifulsoup4>=4.12.0
build>=1.0.0
cachetools>=5.3.0
certifi>=2024.2.0
chardet>=5.2.0
charset-normalizer>=3.3.0
# chroma-hnswlib>=0.7.0  # Requires C++ build tools, using chromadb's built-in HNSW instead
chromadb>=0.4.0
click>=8.1.0
coloredlogs>=15.0.0
dataclasses-json>=0.6.0
dataclasses-json-speakeasy>=0.5.0
Deprecated>=1.2.0
distro>=1.9.0
emoji>=2.10.0
exceptiongroup>=1.2.0
fastapi>=0.109.0
filelock>=3.13.0
filetype>=1.2.0
flatbuffers>=23.5.0
frozenlist>=1.4.0
fsspec>=2024.2.0
google-auth>=2.27.0
googleapis-common-protos>=1.62.0
greenlet>=3.0.0
grpcio>=1.60.0
h11>=0.14.0
httpcore>=1.0.0
httptools>=0.6.0
httpx>=0.26.0
huggingface-hub>=0.20.0
humanfriendly>=10.0
idna>=3.6
importlib-metadata>=6.11.0
importlib-resources>=6.1.0
joblib>=1.3.0
jsonpatch>=1.33
jsonpath-python>=1.0.0
jsonpointer>=2.4
kubernetes>=29.0.0
langchain>=0.1.0
langchain-community>=0.0.19
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-ollama>=0.3.0
langgraph
sentence-transformers
tqdm
langdetect>=1.0.0
langsmith>=0.0.87
lxml>=5.1.0
Markdown>=3.5.0
marshmallow>=3.20.0
mmh3>=4.1.0
monotonic>=1.6
mpmath>=1.3.0
multidict>=6.0.0
mypy-extensions>=1.0.0
nltk>=3.8.0
numpy>=2.1.0
oauthlib>=3.2.0
onnxruntime>=1.17.0
openai>=1.12.0
opentelemetry-api>=1.22.0
opentelemetry-exporter-otlp-proto-common>=1.22.0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
opentelemetry-instrumentation>=0.43b0
opentelemetry-instrumentation-asgi>=0.43b0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-proto>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-semantic-conventions>=0.43b0
opentelemetry-util-http>=0.43b0
overrides>=7.7.0
packaging>=23.2
posthog>=3.4.0
protobuf>=4.25.0
pulsar-client>=3.4.0
pyasn1>=0.5.0
pyasn1-modules>=0.3.0
pydantic>=2.6.0
pydantic_core>=2.16.0
PyPika>=0.48.0
pyproject_hooks>=1.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
python-iso639>=2024.2.0
python-magic>=0.4.0
PyYAML>=6.0.0
rapidfuzz>=3.6.0
regex>=2023.12.0
requests>=2.31.0
requests-oauthlib>=1.3.0
rsa>=4.9
six>=1.16.0
sniffio>=1.3.0
soupsieve>=2.5
SQLAlchemy>=2.0.0
starlette>=0.36.0
sympy>=1.12
tabulate>=0.9.0
tenacity>=8.2.0
tiktoken>=0.5.0
tokenizers>=0.15.0
tomli>=2.0.0
tqdm>=4.66.0
typer>=0.9.0
typing-inspect>=0.9.0
typing_extensions>=4.9.0
unstructured>=0.12.0
unstructured-client>=0.18.0
urllib3>=2.2.0
uvicorn>=0.27.0
# uvloop>=0.19.0  # Not supported on Windows
watchfiles>=0.21.0
websocket-client>=1.7.0
websockets>=12.0
wrapt>=1.16.0
yarl>=1.9.0
zipp>=3.17.0